refactor: rewrite app for using .mjs extension
