+++ package.json
@@ -3,9 +3,8 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./src/app.js"
+    "start": "node ./app.js"
   },
-  "type": "module",
   "dependencies": {
     "axios": "^1.7.2",
     "cookie-parser": "~1.4.4",
