+++ src/backend/controllers/payment-controller.mjs
@@ -1,4 +1,4 @@
-import PaymentService from '../services/payment-service.js';
+import PaymentService from '../services/payment-service.mjs';
 
 class PaymentGatewayController {
