+++ src/backend/utils/validation-utils.mjs
