+++ src/backend/validation/payment-validator.mjs
@@ -1,4 +1,4 @@
-import {checkEmail, checkIsPositiveNumber, isString, notEmpty} from '../utils/validation-utils.js';
+import {checkEmail, checkIsPositiveNumber, isString, notEmpty} from '../utils/validation-utils.mjs';
 
 export const initializePayment = [
     notEmpty(['amount', 'title_service', 'email']),
